Replace `connection.py` with `dbus_connection.py`
